#5 Change usage of entityIdGenerator, refactor test
