+++ backend/test/modules/doubles-tournament/core/application/TestCreateTournamentWithTeamsModule.ts
@@ -1,7 +1,10 @@
 import { TestModuleCore, testModuleCore } from '../../../../test-support/shared/core/TestModuleCore';
 import { DoublesTournamentModuleCore } from '../../../../../src/modules/doubles-tournament/core/DoublesTournamentModuleCore';
+import { EntityIdGeneratorStub } from '../../../../test-support/shared/core/EntityIdGeneratorStub';
 
 export function testCreateTournamentWithTeamsModule(currentTime: Date): TestModuleCore {
-  //TODO it must use some kind of repository for test to work properly
-  return testModuleCore((commandBus, eventBus, queryBus) => DoublesTournamentModuleCore(eventBus, commandBus, () => currentTime));
+  const entityGen = EntityIdGeneratorStub('TeamId');
+  return testModuleCore((commandBus, eventBus, queryBus) =>
+    DoublesTournamentModuleCore(eventBus, commandBus, () => currentTime, entityGen),
+  );
 }
